Jl777 (#1315)
* Remove voutsum check

* Teach RPC interface about dpow-enabled minconfs (#1231)

* Make minconfs dpow-aware in z_listunspent + z_listreceivedbyaddress

* Add dpow-related test files to test suite

* Add dpow simulation to regtest every 7 blocks

* Fix compiler errors

* Fix link error

* Fix stdout spam when running regtests

* Dpowminconfs for listreceivedbyaddress

* dpowconfs tests

* Start adding specific tests for dpowminconfs in listreceivedbyaddress

* Get dpowminconfs tests for listreceivedbyaddress working

* Add dpowminconfs to getreceivedbyaddress + listunspent

* Add test for listtransactions + getreceivedbyaddress support

* Reliably passing dpowminconf tests. We only check for notarized-ness now, not exact confirmation numbers, to avoid race conditions

* Poll for the expected notarization info before running further tests; add support for getbalance

* Migrate tx_height() to a place where asyncrpcoperation_sendmany.cpp can use it

* fix

* Teach GetFilteredNotes about dpowconfs

Many RPCs rely on this internal function, which now correctly uses
dpowconfs to filter by the minconf/maxconf parameters.

* Fix sendmany when using non-default minconf

* inline seems to make things happy

* cleanup

* Add some code to test z_sendmany, which points out https://github.com/jl777/komodo/issues/1247

* try this

* Use already calculated value of dpowconfs instead of calculating it again

* Cleanup .pack file

* Remove

* Remove .pack

* Disable passkeys

* Rvalidate

* Syntax

* Allow overwrite by same pub33

* Tx

* Declare variables

* Allow replacement handle

* Grandfather existing handles

* Test

* Handleinfo

* Char *

* Begin

* Add mutex

* CCaddr

* Casts for windows

* +debugs

* Syntax

* Item

* Skeet

* +print

* Error check things

* +prints

* -sleep

* Brute force inventory check

* Revert

* num_packitems

* Log file

* Test

* ABC

* Test

* Add help docs for all -ac_* params of komodod (#1313)

* Remove myAddress from roc

* +print

* Test

* Test

* Leave pack all

* Dont discard unless last o_count

* Prevent pack corruption

* -ddebugs

* Merge branch 'FSM' into jl777

# Conflicts:
#	src/cc/rogue/main.c

+++ src/wallet/walletdb.cpp
@@ -965,7 +965,7 @@ DBErrors CWalletDB::LoadWallet(CWallet* pwallet)
                 reAdded++;
             }
         }
-        fprintf(stderr, "Cleared %lu corrupted transactions from wallet. Readded %i known transactions.\n",deadTxns.size(),reAdded);
+        fprintf(stderr, "Cleared %lu corrupted transactions from wallet. Readded %i known transactions.\n",(long)deadTxns.size(),reAdded);
         deadTxns.clear();
     }
